Move ReflectionInformation into ssz.
